refactor: 우상자 찾기를 클래스로 분리
