+++ src/main/kotlin/game/domain/WinnerFinder.kt
@@ -1,0 +1,5 @@
+package game.domain
+
+fun interface WinnerFinder {
+    fun findWinner(cars: List<Car>): List<String>
+}
